Struct nautilus_network::http::HttpClient

pub struct HttpClient { /* private fields */ }

Provides a high-performance HttpClient for HTTP requests.

The client is backed by a hyper Client which keeps connections alive and can be cloned cheaply. The client also has a list of header fields to extract from the response.

The client returns an HttpResponse. The client filters only the key value for the give header_keys.

Implementations

impl HttpClient

pub fn new(header_keys: Vec<String>) -> Self

pub fn request<'py>( slf: PyRef<'_, Self>, method_str: String, url: String, headers: HashMap<String, String>, body: Option<&'py PyBytes>, py: Python<'py> ) -> PyResult<&'py PyAny>

pub fn get<'py>( slf: PyRef<'_, Self>, url: String, headers: HashMap<String, String>, body: Option<&'py PyBytes>, py: Python<'py> ) -> PyResult<&'py PyAny>

pub fn post<'py>( slf: PyRef<'_, Self>, url: String, headers: HashMap<String, String>, body: Option<&'py PyBytes>, py: Python<'py> ) -> PyResult<&'py PyAny>

pub fn patch<'py>( slf: PyRef<'_, Self>, url: String, headers: HashMap<String, String>, body: Option<&'py PyBytes>, py: Python<'py> ) -> PyResult<&'py PyAny>

pub fn delete<'py>( slf: PyRef<'_, Self>, url: String, headers: HashMap<String, String>, body: Option<&'py PyBytes>, py: Python<'py> ) -> PyResult<&'py PyAny>

impl HttpClient

pub async fn send_request( &self, method: Method, url: String, headers: HashMap<String, String>, body: Option<Vec<u8>> ) -> Result<HttpResponse, Box<dyn Error + Send + Sync>>

pub async fn to_response( &self, res: Response<Body> ) -> Result<HttpResponse, Box<dyn Error + Send + Sync>>