nautilus_network::ratelimiter::clock

Trait Reference

pub trait Reference:
    Sized
    + Add<Nanos, Output = Self>
    + PartialEq
    + Eq
    + Ord
    + Copy
    + Clone
    + Send
    + Sync
    + Debug {
    // Required methods
    fn duration_since(&self, earlier: Self) -> Nanos;
    fn saturating_sub(&self, duration: Nanos) -> Self;
}

A measurement from a clock.

Required Methods

fn duration_since(&self, earlier: Self) -> Nanos

Determines the time that separates two measurements of a clock. Implementations of this must perform a saturating subtraction - if the earlier timestamp should be later, duration_since must return the zero duration.

fn saturating_sub(&self, duration: Nanos) -> Self

Returns a reference point that lies at most duration in the past from the current reference. If an underflow should occur, returns the current reference.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.