Struct nautilus_network::socket::SocketClient

pub struct SocketClient { /* private fields */ }

Creates a TcpStream with the server

The stream can be encrypted with TLS or Plain. The stream is split into read and write ends.

The read end is passed to task that keeps receiving messages from the server and passing them to a handler.

The write end is wrapped in an Arc Mutex and used to send messages or heart beats

The heartbeat is optional and can be configured with an interval and message.

The client uses a suffix to separate messages on the byte stream. It is appended to all sent messages and heartbeats. It is also used the split the received byte stream.

Implementations

impl SocketClient

pub async fn connect_url( url: &str, handler: PyObject, mode: Mode, suffix: Vec<u8>, heartbeat: Option<(u64, Vec<u8>)> ) -> Result<Self>

pub async fn tls_connect_with_server( url: &str, mode: Mode ) -> (ReadHalf<MaybeTlsStream<TcpStream>>, WriteHalf<MaybeTlsStream<TcpStream>>)

pub fn spawn_read_task( reader: ReadHalf<MaybeTlsStream<TcpStream>>, handler: PyObject, suffix: Vec<u8> ) -> JoinHandle<()>

pub fn spawn_heartbeat_task( heartbeat: Option<(u64, Vec<u8>)>, writer: Arc<Mutex<WriteHalf<MaybeTlsStream<TcpStream>>>>, suffix: Vec<u8> ) -> Option<JoinHandle<()>>

Optionally spawn a hearbeat task to periodically ping the server.

pub async fn shutdown(&mut self)

Shutdown read task and the connection.

The client must be explicitly shutdown before dropping otherwise the connection might still be alive for some time before terminating. Closing the connection is an async call which cannot be done by the drop method so it must be done explicitly.

pub async fn send_bytes(&mut self, data: &[u8])

pub fn is_alive(&self) -> bool

Checks if the client is still connected.